+++ apps/ensapi/package.json
@@ -28,7 +28,7 @@
     "@ensnode/ponder-subgraph": "workspace:*",
     "@hono/node-server": "^1.19.5",
     "@hono/otel": "^0.2.2",
-    "@hono/zod-openapi": "1.0.0-beta.1",
+    "@hono/zod-openapi": "1.2.1",
     "@namehash/ens-referrals": "workspace:*",
     "@opentelemetry/api": "^1.9.0",
     "@opentelemetry/core": "^2.0.1",
