+++ apps/ensapi/src/handlers/name-tokens-api.ts
@@ -107,7 +107,7 @@ app.openapi(nameTokensGetMeta, async (c): Promise<any> => {
       return c.json(
         serializeNameTokensResponse(
           makeNameTokensNotIndexedResponse(
-            `This ENSNode instance has not been configured to index tokens for the requested name: '${name}`,
+            `This ENSNode instance has not been configured to index tokens for the requested name: '${name}'`,
           ),
         ),
         404,
